requirements.txt
Django==5.1.3
djangorestframework==3.15.2
django-filter==24.3
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.94.1
drf-spectacular==0.27.2
PyJWT==2.9.0
sqlparse==0.5.1